@@ -2,7 +2,7 @@
 #
 # The admin can manage and change the cms pages content from the backEnd
 #
-class ContentManagementPage < ActiveRecord::Base
+class ContentManagementPage < ApplicationRecord
 
   # For use search method
   extend Search
@@ -11,5 +11,7 @@ class ContentManagementPage < ActiveRecord::Base
   validates :page_title, :content, :presence => true
 
   translates :content, :brief, :page_title
-
+  attribute :content
+  attribute :brief
+  attribute :page_title
 end
