@@ -2,7 +2,7 @@
 #
 # The admin can add posts for every category
 #
-class Post < ActiveRecord::Base
+class Post < ApplicationRecord
 
   ## DB relations
   belongs_to :person
@@ -11,6 +11,20 @@ class Post < ActiveRecord::Base
   has_and_belongs_to_many :social_media_platforms, through: :posts_social_media_platforms
 
   translates :title, :description, :intro, :tags, :side_image_caption, :quote_author, :quote_text, :twitter_text, :news_source, :image_author_name, :pdf_file_name, :pdf_content_type, :pdf_file_size, :pdf_updated_at, :fallbacks_for_empty_translations => true
+  attribute :title
+  attribute :description
+  attribute :intro
+  attribute :tags
+  attribute :side_image_caption
+  attribute :quote_author
+  attribute :quote_text
+  attribute :twitter_text
+  attribute :news_source
+  attribute :image_author_name
+  attribute :pdf_file_name
+  attribute :pdf_content_type
+  attribute :pdf_file_size
+  attribute :pdf_updated_at
 
   # For use search method
   extend Search
