@@ -2,7 +2,7 @@
 #
 # The every post can have quote 
 #
-class Quote < ActiveRecord::Base
+class Quote < ApplicationRecord
 
   ## DB relations
   has_and_belongs_to_many :posts_themes, through: :quotes_posts_themes
@@ -18,5 +18,7 @@ class Quote < ActiveRecord::Base
   validates :author,:font_size,:quote, :presence => true
 
   translates :author, :quote, :twitter_text
-
+  attribute :author
+  attribute :quote
+  attribute :twitter_text
 end
